Apply review comments
+ Move frame execution context to frame.go

+ Remove the unnecessary Errorf in Frame.document

+ Make the test robust
